refactor: Migrate TicketAttachmentDownload to singleton service for improved management
- Refactored the TicketAttachmentDownloadQueueHandler to a TicketAttachmentDownloadService.
- The TicketAttachmentDownloadService is now registered as a singleton in BusinessDependency.cs for dependency injection, ensuring a single instance is used throughout the application.
