+++ src/Modmail.NET.Web.Blazor/Dependency/BusinessDependency.cs
@@ -3,6 +3,7 @@ using Modmail.NET.Database;
 using Modmail.NET.Database.Triggers;
 using Modmail.NET.Language;
 using Modmail.NET.Queues;
+using Modmail.NET.Services;
 using Modmail.NET.Static;
 using Modmail.NET.Web.Blazor.Services;
 using Serilog;
@@ -26,6 +27,7 @@ public static class BusinessDependency
     builder.Services.AddSingleton<ModmailBot>();
     builder.Services.AddSingleton<ModmailEventHandlers>();
     builder.Services.AddSingleton<TicketMessageQueue>();
+    builder.Services.AddSingleton<TicketAttachmentDownloadService>();
     builder.Services.AddHostedService<ModmailHostedService>();
 
     builder.Services.AddDbContextFactory<ModmailDbContext>(options => {
